fix(app-blocks): move catalog test+helper files out of src/pages (next build ApiRouteConfig)
Next type-checks every file under src/pages/** as an API route, so a
non-route file there fails the ApiRouteConfig constraint at `next build`
(tsc --noEmit does not catch it). The PR-preview build failed with:
  Type 'typeof import(".../blocks/__tests__/images-endpoint.test")'
  does not satisfy the constraint 'ApiRouteConfig'.

Relocate the 6 offending files out of src/pages (pure move + import
repointing, no behavior change):
  - catalog-maturity.ts -> src/server/utils/block-catalog-maturity.ts
    (Prisma-free helper; matches the server/utils convention; importers
     in blocks/images.ts + blocks/models.ts repointed to the ~/ path)
  - 5 test files -> src/tests/api/v1/{blocks,images,models}/...
    (matches the existing src/tests/api/blocks/*.test.ts convention)

All relative imports (../models, ../images, ../index, ../catalog-maturity)
switched to ~/-aliased absolute imports. No .ts file under src/pages/api/**
is now a non-route. 33 relocated tests green; typecheck clean; next build
TypeScript/compile phase passes with no ApiRouteConfig error.
