+++ src/tests/api/v1/models/index-refactor.test.ts
@@ -75,7 +75,7 @@ function fakeRes() {
 }
 
 async function invoke(query: Record<string, unknown>) {
-  const mod = await import('../index');
+  const mod = await import('~/pages/api/v1/models/index');
   const handler = mod.default as any;
   const req = {
     method: 'GET',
